Modifications for DAT proposal
Some values are hard-coded, and all of the fialkov models are computed in the run_batch.py loop.

The main area to confirm is the uv coverage calculations. Ideally, we would use a gridder that takes into
account the antenna gain pattern by using a convolution kernel in uv space.